+++ client/src/lib/state/dns.js
@@ -14,6 +14,9 @@ const {
   isEmpty,
   isInRange,
   revision,
+  isWholeNumber,
+  nestedSplat,
+  contains,
 } = require("lazy-z");
 const {
   shouldDisableComponentSave,
@@ -24,6 +27,7 @@ const {
   resourceGroupsField,
   unconditionalInvalidText,
   vpcGroups,
+  subnetMultiSelect,
 } = require("./utils");
 const {
   invalidName,
@@ -227,6 +231,24 @@ function dnsResolverDelete(config, stateData, componentProps) {
 }
 
 /**
+ * hide when record is not SRV
+ * @param {*} stateData
+ * @returns {boolean} true if should be hidden
+ */
+function hideWhenRecordNotSrv(stateData) {
+  return stateData.type !== "SRV";
+}
+
+/**
+ * hide when not targeting vs
+ * @param {*} stateData
+ * @returns {boolean} true when should be hidden
+ */
+function hideWhenNotVsi(stateData) {
+  return stateData.use_vsi !== true;
+}
+
+/**
  * init dns store
  * @param {*} store
  */
@@ -243,18 +265,21 @@ function initDnsStore(store) {
     ),
     schema: {
       name: {
+        size: "small",
         default: "",
         invalid: invalidName("dns"),
         invalidText: invalidNameText("dns"),
       },
       plan: {
+        size: "small",
+        type: "select",
         default: "free",
         groups: ["Free", "Standard"],
         invalidText: selectInvalidText("plan"),
         onInputChange: kebabCaseInput("plan"),
         onRender: titleCaseRender("plan"),
       },
-      resource_group: resourceGroupsField(),
+      resource_group: resourceGroupsField(true),
     },
     subComponents: {
       zones: {
@@ -268,12 +293,31 @@ function initDnsStore(store) {
         ),
         schema: {
           name: {
+            size: "small",
             default: "",
-            invalid: invalidDnsZoneName,
+            invalid: function (stateData, componentProps) {
+              let allZoneNames = [];
+              componentProps.craig.store.json.dns.forEach((instance) => {
+                instance.zones.forEach((zone) => {
+                  if (
+                    (componentProps.data &&
+                      componentProps.data.name !== zone.name) ||
+                    componentProps.isModal
+                  ) {
+                    allZoneNames.push(zone.name);
+                  }
+                });
+              });
+              return contains(allZoneNames, stateData.name)
+                ? true
+                : invalidDnsZoneName(stateData, componentProps);
+            },
             invalidText: invalidNameText("zones"),
           },
           vpcs: {
-            type: "multiselct",
+            size: "small",
+            labelText: "VPC Networks",
+            type: "multiselect",
             default: [],
             invalid: function (stateData) {
               return (
@@ -284,6 +328,7 @@ function initDnsStore(store) {
             groups: vpcGroups,
           },
           label: {
+            size: "small",
             default: "",
             invalid: fieldIsNullOrEmptyString("label"),
             invalidText: unconditionalInvalidText(
@@ -291,6 +336,7 @@ function initDnsStore(store) {
             ),
           },
           description: {
+            labelText: "Description",
             type: "textArea",
             default: "",
             invalid: function (stateData) {
@@ -299,6 +345,7 @@ function initDnsStore(store) {
             invalidText: unconditionalInvalidText(
               "Invalid description. Must match the regex expression /^[a-zA-Z0-9]+$/."
             ),
+            placeholder: "(Optional) DNS Zone Description",
           },
         },
       },
@@ -318,17 +365,75 @@ function initDnsStore(store) {
             "priority",
             "service",
             "weight",
+            "vpc",
+            "vsi",
+            "ttl",
           ],
           "dns",
           "records"
         ),
         schema: {
+          use_vsi: {
+            type: "toggle",
+            labelText: "Use VSI IP Address",
+            default: false,
+            tooltip: {
+              content:
+                "Use the IP address of a primary network interface for a VPC Virtual Server as the resource data for this record",
+            },
+          },
+          vpc: {
+            size: "small",
+            labelText: "VPC",
+            default: "",
+            type: "select",
+            groups: vpcGroups,
+            invalidText: selectInvalidText("VPC"),
+            invalid: function (stateData) {
+              return stateData.use_vsi
+                ? isNullOrEmptyString(stateData.vpc, true)
+                : false;
+            },
+            hideWhen: hideWhenNotVsi,
+          },
+          vsi: {
+            size: "small",
+            labelText: "Virtual Server Instance",
+            default: "",
+            type: "select",
+            invalid: function (stateData) {
+              return stateData.use_vsi
+                ? isNullOrEmptyString(stateData.vsi, true)
+                : false;
+            },
+            invalidText: selectInvalidText("VSI"),
+            groups: function (stateData, componentProps) {
+              let allVsi = componentProps.craig.store.json.vsi.filter((vsi) => {
+                if (vsi.vpc === stateData.vpc) {
+                  return vsi;
+                }
+              });
+              let allVsiNames = [];
+              allVsi.forEach((deployment) => {
+                for (let i = 0; i < deployment.subnets.length; i++) {
+                  for (let j = 0; j < deployment.vsi_per_subnet; j++) {
+                    allVsiNames.push(`${deployment.name}-${i + 1}-${j + 1}`);
+                  }
+                }
+              });
+              return allVsiNames;
+            },
+            hideWhen: hideWhenNotVsi,
+          },
           name: {
+            size: "small",
             default: "",
             invalid: invalidName("records"),
             invalidText: invalidNameText("records"),
           },
           dns_zone: {
+            size: "small",
+            labelText: "DNS Zone",
             type: "select",
             default: "",
             invalid: fieldIsNullOrEmptyString("dns_zone"),
@@ -344,14 +449,41 @@ function initDnsStore(store) {
             },
           },
           rdata: {
+            size: "small",
+            labelText: "Resource Data",
             default: "",
             invalid: fieldIsNullOrEmptyString("rdata"),
+            invalidText: unconditionalInvalidText(
+              "Resource Data cannot be null or empty string."
+            ),
+            hideWhen: function (stateData) {
+              return stateData.use_vsi === true;
+            },
+          },
+          ttl: {
+            size: "small",
+            labelText: "Time To Live (Seconds)",
+            default: "",
+            invalid: function (stateData) {
+              let ttl = parseFloat(stateData.ttl);
+              return isNullOrEmptyString(stateData.ttl, true)
+                ? false
+                : !isWholeNumber(ttl) || !isInRange(ttl, 300, 2147483647);
+            },
+            invalidText: unconditionalInvalidText(
+              "Enter a whole number between 300 and 2147483647"
+            ),
           },
           type: {
+            size: "small",
+            type: "select",
             default: "",
             invalid: fieldIsNullOrEmptyString("type"),
+            invalidText: selectInvalidText("DNS record type"),
+            groups: ["A", "AAAA", "CNAME", "PTR", "TXT", "MX", "SRV"],
           },
           preference: {
+            size: "small",
             default: "",
             invalid: function (stateData) {
               if (stateData.type === "MX") {
@@ -360,8 +492,15 @@ function initDnsStore(store) {
                 } else return true;
               } else return false;
             },
+            invalidText: unconditionalInvalidText(
+              "Must be a whole number within range 0 and 65535"
+            ),
+            hideWhen: function (stateData) {
+              return stateData.type !== "MX";
+            },
           },
           port: {
+            size: "small",
             default: "",
             invalid: function (stateData) {
               if (stateData.type === "SRV") {
@@ -370,8 +509,14 @@ function initDnsStore(store) {
                 } else return true;
               } else return false;
             },
+            invalidText: unconditionalInvalidText(
+              "Must be a whole number between 1 and 65535"
+            ),
+            hideWhen: hideWhenRecordNotSrv,
           },
           protocol: {
+            size: "small",
+            type: "select",
             default: "",
             invalid: function (stateData) {
               if (stateData.type === "SRV") {
@@ -381,8 +526,12 @@ function initDnsStore(store) {
                 );
               } else return false;
             },
+            invalidText: selectInvalidText("protocol"),
+            groups: ["TCP", "UDP"],
+            hideWhen: hideWhenRecordNotSrv,
           },
           priority: {
+            size: "small",
             default: "",
             invalid: function (stateData) {
               if (stateData.type === "SRV") {
@@ -391,8 +540,13 @@ function initDnsStore(store) {
                 } else return true;
               } else return false;
             },
+            invalidText: unconditionalInvalidText(
+              "Must be a whole number between 0 and 65535"
+            ),
+            hideWhen: hideWhenRecordNotSrv,
           },
           service: {
+            size: "small",
             default: "",
             invalid: function (stateData) {
               if (stateData.type === "SRV") {
@@ -404,8 +558,13 @@ function initDnsStore(store) {
                 } else return true;
               } else return false;
             },
+            invalidText: unconditionalInvalidText(
+              "Service must start with a '_'."
+            ),
+            hideWhen: hideWhenRecordNotSrv,
           },
           weight: {
+            size: "small",
             default: "",
             invalid: function (stateData) {
               if (stateData.type === "SRV") {
@@ -414,6 +573,10 @@ function initDnsStore(store) {
                 } else return true;
               } else return false;
             },
+            invalidText: unconditionalInvalidText(
+              "Must be a whole number between 0 and 65535"
+            ),
+            hideWhen: hideWhenRecordNotSrv,
           },
         },
       },
@@ -428,25 +591,39 @@ function initDnsStore(store) {
         ),
         schema: {
           name: {
+            size: "small",
             default: "",
             invalid: invalidName("custom_resolvers"),
             invalidText: invalidNameText("custom_resolvers"),
           },
           vpc: {
+            size: "small",
+            type: "select",
+            labelText: "VPC",
             default: "",
             invalid: fieldIsNullOrEmptyString("vpc"),
+            invalidText: selectInvalidText("VPC"),
+            groups: vpcGroups,
           },
-          subnets: {
-            default: "",
+          subnets: subnetMultiSelect({
             invalid: function (stateData) {
-              return isEmpty(stateData.subnets);
+              return stateData.subnets.length > 3;
             },
-          },
+            invalidText: "Select between 1 and 3 subnets",
+          }),
           description: {
+            labelText: "Description",
+            type: "textArea",
             default: "",
             invalid: function (stateData) {
               return invalidDescription(stateData.description);
             },
+            invalidText: unconditionalInvalidText(
+              "Invalid description. Must match the regex expression /^[a-zA-Z0-9]+$/."
+            ),
+            placeholder: unconditionalInvalidText(
+              "(Optional) Custom Resolver Description"
+            ),
           },
         },
       },