+++ unit-tests/state/dns.test.js
@@ -220,6 +220,33 @@ describe("dns", () => {
         "it should delete zone"
       );
     });
+    it("should disable save when zone name duplicate in modal", () => {
+      state.dns.zones.create(
+        {
+          name: "zone",
+          vpcs: [],
+        },
+        {
+          innerFormProps: { arrayParentName: "dev" },
+        }
+      );
+      state.dns.zones.create(
+        {
+          name: "zone2",
+          vpcs: [],
+        },
+        {
+          innerFormProps: { arrayParentName: "dev" },
+        }
+      );
+      assert.isTrue(
+        state.dns.zones.name.invalid(
+          { name: "zone" },
+          { isModal: true, craig: state }
+        ),
+        "it should be invalid"
+      );
+    });
   });
   describe("dns.records", () => {
     let state;
@@ -310,6 +337,91 @@ describe("dns", () => {
           ),
           [],
           "it should return correct groups"
+        );
+      });
+      it("should hide preference when type is not MX", () => {
+        assert.isTrue(
+          state.dns.records.preference.hideWhen({ type: "A" }),
+          "it should be hidden"
+        );
+      });
+      it("should hide port when type is not SRV", () => {
+        assert.isTrue(
+          state.dns.records.port.hideWhen({ type: "A" }),
+          "it should be hidden"
+        );
+      });
+      it("should have invalid ttl when not a whole number", () => {
+        assert.isTrue(
+          state.dns.records.ttl.invalid({ ttl: "1.2" }),
+          "it should be invalid"
+        );
+      });
+      it("should have invalid ttl when not in range", () => {
+        assert.isTrue(
+          state.dns.records.ttl.invalid({ ttl: "2" }),
+          "it should be invalid"
+        );
+      });
+      it("should return false for vpc invalid if not using vsi", () => {
+        assert.deepEqual(
+          state.dns.records.vpc.invalid({}),
+          false,
+          "it should be false"
+        );
+      });
+      it("should return true for vpc invalid if using vsi and no vpc selected", () => {
+        assert.deepEqual(
+          state.dns.records.vpc.invalid({ use_vsi: true, vpc: "" }),
+          true,
+          "it should be true"
+        );
+      });
+      it("should return true for vsi invalid if using vsi and no vsi selected", () => {
+        assert.deepEqual(
+          state.dns.records.vsi.invalid({ use_vsi: true, vsi: "" }),
+          true,
+          "it should be true"
+        );
+      });
+      it("should return false for vsi invalid if using vsi and no vpc selected", () => {
+        assert.deepEqual(
+          state.dns.records.vsi.invalid({ use_vsi: false, vsi: "" }),
+          false,
+          "it should be true"
+        );
+      });
+      it("should return a list of vsi based on vpc", () => {
+        assert.deepEqual(
+          state.dns.records.vsi.groups({ vpc: "management" }, { craig: state }),
+          [
+            "management-server-1-1",
+            "management-server-1-2",
+            "management-server-2-1",
+            "management-server-2-2",
+            "management-server-3-1",
+            "management-server-3-2",
+          ],
+          "it should return list of vsi"
+        );
+      });
+      it("should return a list of vsi based on vpc when none selected", () => {
+        assert.deepEqual(
+          state.dns.records.vsi.groups({ vpc: "" }, { craig: state }),
+          [],
+          "it should return list of vsi"
+        );
+      });
+      it("should hide rdata when using vsi", () => {
+        assert.isTrue(
+          state.dns.records.rdata.hideWhen({ use_vsi: true }),
+          "it should be hidden"
+        );
+      });
+      it("should hide vsi when not using vsi", () => {
+        assert.isTrue(
+          state.dns.records.vsi.hideWhen({ use_vsi: false }),
+          "it should be hidden"
         );
       });
     });